// src/app/interfaces/aplicaciones.interface.ts
export interface AplicacionesInterface {
    application:        number;
    application_Father: number;
    description:        string;
    observacion_1:      string;
    //Colores
    colorBackground:    string;    //Color de fonde de la app.
    colorFontNotSelect: string; //Color de fuente no seleccionada.
    colorFontSelect:    string;   //Color de fuente seleccionada
    colorSelected:      string;      //Color de objeto seleccionado
    colorMargenSelect:  string;  //Color de barra laterel al seleccionar onjeto.
}